requirements.txt
alembic==1.5.7
click==7.1.2
Flask-Migrate==2.7.0
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.3
Mako==1.1.4
MarkupSafe==1.1.1
pip==21.0.1
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
setuptools==54.1.2
six==1.15.0
Werkzeug==1.0.1
WTForms==2.3.3
